+++ lib/src/public/base/stream_bloc.dart
@@ -7,7 +7,6 @@ import 'package:stream_bloc/src/public/interfaces/stream_bloc.dart';
 /// as an output.
 ///
 /// Almost identical to pre-`7.2.0` Bloc from the `bloc` package.
-
 abstract class StreamBloc<Event extends Object?,
         State extends Object?> = StreamBlocBase<Event, State>
     with