+++ lib/src/public/base/stream_bloc_observer.dart
@@ -3,19 +3,27 @@ import 'dart:async';
 import 'package:bloc/bloc.dart';
 import 'package:meta/meta.dart';
 
-import 'package:stream_bloc/src/public/interfaces/stream_bloc.dart';
-
+/// Allows observing certain lifecycle stages of corresponding interfaces and
+/// provides a mechanism of injecting itself through [Zone]s.
+///
+/// Combines in itself a [BlocObserver] with interface arguments and a
+/// [BlocOverrides] that allows overriding only an observer.
 abstract class StreamBlocObserver {
-  static const _tag = #stream_bloc_observer_tag;
+  static const Object _tag = Object();
 
+  /// Runs `body` in a fresh [Zone] using the provided `observer`.
+  ///
+  /// Similar to [BlocOverrides.runZoned].
   static R inject<R>(
     StreamBlocObserver observer,
     R Function() body,
   ) =>
       runZoned(body, zoneValues: {_tag: observer});
 
+  /// Returns the current [StreamBlocObserver] instance.
   static StreamBlocObserver? get current {
     Zone? currentZone = Zone.current;
+
     while (currentZone != null) {
       final Object? value = currentZone[_tag];
       if (value is StreamBlocObserver) return value;
@@ -29,20 +37,20 @@ abstract class StreamBlocObserver {
   /// instance is created.
   @protected
   @mustCallSuper
-  void onCreate(IStreamBloc bloc) {}
+  void onCreate(Closable closable) {}
 
   /// Called whenever an [event] is `added` to any [bloc] with the given [bloc]
   /// and [event].
   @protected
   @mustCallSuper
-  void onEvent(IStreamBloc bloc, Object? event) {}
+  void onEvent(BlocEventSink eventSink, Object? event) {}
 
   /// Called whenever a [Change] occurs in any [bloc]
   /// A [change] occurs when a new state is emitted.
   /// [onChange] is called before a bloc's state has been updated.
   @protected
   @mustCallSuper
-  void onChange(IStreamBloc bloc, Change change) {}
+  void onChange(StateStreamable stateStreamable, Change change) {}
 
   /// Called whenever a transition occurs in any [bloc] with the given [bloc]
   /// and [transition].
@@ -51,14 +59,14 @@ abstract class StreamBlocObserver {
   /// [onTransition] is called before a [bloc]'s state has been updated.
   @protected
   @mustCallSuper
-  void onTransition(IStreamBloc bloc, Transition transition) {}
+  void onTransition(BlocEventSink bloc, Transition transition) {}
 
   /// Called whenever an [error] is thrown in any [Bloc] or [Cubit].
   /// The [stackTrace] argument may be [StackTrace.empty] if an error
   /// was received without a stack trace.
   @protected
   @mustCallSuper
-  void onError(IStreamBloc bloc, Object error, StackTrace stackTrace) {}
+  void onError(ErrorSink errorSink, Object error, StackTrace stackTrace) {}
 
   /// Called whenever a [Bloc] is closed.
   /// [onClose] is called just before the [Bloc] is closed
@@ -66,5 +74,5 @@ abstract class StreamBlocObserver {
   /// emit new states.
   @protected
   @mustCallSuper
-  void onClose(IStreamBloc bloc) {}
+  void onClose(Closable closable) {}
 }