+++ lib/src/public/interfaces/stream_bloc.dart
@@ -1,12 +1,11 @@
-import 'package:bloc/bloc.dart';
 import 'package:stream_bloc/src/public/interfaces/stream_bloc_hooks.dart';
 import 'package:stream_bloc/src/public/interfaces/stream_bloc_mapper.dart';
 import 'package:stream_bloc/src/public/interfaces/stream_bloc_transformers.dart';
 
+/// An interface that combines [StreamBlocMapper], [StreamBlocTransformers] and
+/// [StreamBlocHooks] and represents all required methods for a `StreamBloc`
 abstract class IStreamBloc<Event extends Object?, State extends Object?>
     implements
-        StateStreamableSource<State>,
-        BlocEventSink<Event>,
         StreamBlocMapper<Event, State>,
         StreamBlocTransformers<Event, State>,
         StreamBlocHooks<Event, State> {}