+++ lib/src/public/interfaces/stream_bloc_mapper.dart
@@ -1,6 +1,11 @@
+import 'package:bloc/bloc.dart';
 import 'package:meta/meta.dart';
 
-abstract class StreamBlocMapper<Event extends Object?, State extends Object?> {
+/// An interface that implement [BlocEventSink] and [StateStreamableSource],
+/// adding a [mapEventToStates] method, a method that describes an asynchronous
+/// "one-to-many" relationship between `events` and `states`.
+abstract class StreamBlocMapper<Event extends Object?, State extends Object?>
+    implements BlocEventSink<Event>, StateStreamableSource<State> {
   /// Must be implemented when a class extends [IStreamBloc].
   /// [mapEventToStates] is called whenever an [event] is [add]ed
   /// and is responsible for converting that [event] into a new [state].