+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.2.0 - [04.02.2022]
+
+- Events stream is made broadcast.
+- Package exports include the `bloc` package
+- Fixed meta dependency, bumping it to the latest version
+- Added missing `on...` arguments to `BlocLifecycleMixin`'s methods
+- `StreamBlocObserver`'s methods expect interfaces
+- Added missing documentation. The package is now 100% documented.
+
 ## 0.1.0 - [06.01.2022]
 
 - Initial version.